Update patch set 3
Patch Set 3:

We are not yet moving to the new names for the serverIDs

this is a change that has to be done across all projects to keep consistency, but for now many teams are used to these names. 

We can always coordinate with Gildas and all the teams to make this name change maybe after the release.

Patch-set: 3
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0